+++ CHANGELOG.md
@@ -1,4 +1,13 @@
 # Changelog
+## [1.0.2] - 2021-07-12
+### Removed
+- scrollOffsetHeight property.
+- init method.
+
+### Added
+- barColor property for coloring bars without styles.
+- offsettingBar instead scrollOffsetHeight. All styles calculate inside component.
+- resize method contains createStyles and moveBar;
 
 ## [1.0.1] - 2020-01-06